requirements.txt
streamlit==1.35.0
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
seaborn==0.13.2
plotly==5.22.0
pdfplumber==0.10.3
scikit-learn==1.4.2
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
